+++ src/Controller/SigninController.php
@@ -2,21 +2,17 @@
 
 namespace App\Controller;
 
+use App\Model\Twig;
 use App\Model\User;
 use App\Model\UserManager;
 use Twig\Environment;
 use Twig\Loader\FilesystemLoader;
 
-class SigninController
+class SigninController extends Twig
 {
-    public function show() : void
+    public function show($filter = null) : void
     {
-        $loader = new FilesystemLoader('src/View');
-        $twig = new Environment($loader, [
-            'cache' => false//'src/tmp',
-        ]);
-
-        echo $twig->render('signin.html.twig');
+        $this->twig('signin.html.twig',['erreur'=>''.$filter.'']);
     }
 
     public function signIn() : void
@@ -26,8 +22,19 @@ class SigninController
             'password' => $_POST['password']
        ]);
         $userManager = new UserManager();
-        $userManager->connectUser($user);
-        $home = new HomeController();
-        $home->show();
+        if (empty($_POST["password"]) || empty($_POST["pseudo"])){
+            $this->show('Veuillez remplir tout les champs');
+        }
+        elseif (!$userManager->checkIfPseudoExist($user)){
+            $this->show('Pseudo incorrect');
+        }
+        elseif (!$userManager->checkPasswordHash($user)){
+            $this->show('Mauvais mot de passe');
+        }
+        else{
+            $userManager->connectUser($user);
+            $home = new HomeController();
+            $home->show();
+        }
     }
 }
